Fixed ShellExecute crash bug
I forgot that i had the same crash bug when using GetOpenFileName().
When a win32 dialog is open and blocking a redy program, it crashes
because the WM_TIMER event handler has a task_yield() in it. The
solution was to temporarily disable the task_yield() call while a dialog
is open. I added that to the ShellExecute wrapper, and now it works fine
even if the security dialog pops up.

Now that i think about it, this could explain why ShellExecute would
crash every once in a while, even when it doesn't need to ask for
permission to run the exe. If it takes a fraction of a second too long
for the function to return, i guess the 0.1 second timer that's running
may call the callback in that amount of time, causing a rare crash.